better search testing
