+++ test-suite/Search_test.go
@@ -1,0 +1,43 @@
+package testSuite
+
+import (
+	"testing"
+
+	testUtils "github.com/replit/upm/test-suite/utils"
+)
+
+func TestSearch(t *testing.T) {
+	for _, bt := range languageBackends {
+		bt.Start(t)
+
+		switch bt.Backend.Name {
+		case "bun":
+			fallthrough
+		case "nodejs-npm":
+			fallthrough
+		case "nodejs-pnpm":
+			fallthrough
+		case "nodejs-yarn":
+			doSearch(bt, []searchTest{
+				{"express", "express"},
+				{"@replit/crosis", "@replit/crosis"},
+				{"@replit", "@replit/crosis"},
+			})
+		}
+	}
+}
+
+type searchTest struct {
+	query    string
+	expected string
+}
+
+func doSearch(t testUtils.BackendT, tests []searchTest) {
+	t.Subtest(t.Backend.Name, func(t testUtils.BackendT) {
+		for _, test := range tests {
+			t.Subtest(test.query, func(t testUtils.BackendT) {
+				t.UpmSearch(test.query, test.expected)
+			})
+		}
+	})
+}
